// src/app/api/generated/model/assign-grant-request.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Permiso adicional a otorgar sobre un vinculo
 */
export interface AssignGrantRequest { 
    /**
     * Codigo del permiso, del catalogo de la matriz
     */
    permissionCode: string;
    /**
     * Motivo declarado del otorgamiento. Queda en la auditoria
     */
    reason: string;
    /**
     * Vencimiento del permiso adicional. Sin el, rige hasta que se de de baja
     */
    validUntil?: string;
}