// src/app/api/generated/model/membership-page-response.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MembershipResponse } from './membership-response';


/**
 * Pagina de colaboradores de la organizacion
 */
export interface MembershipPageResponse { 
    /**
     * Vinculos de esta pagina
     */
    content?: Array<MembershipResponse>;
    /**
     * Numero de pagina devuelta, base cero
     */
    page?: number;
    /**
     * Cantidad de elementos por pagina
     */
    size?: number;
    /**
     * Cantidad total de vinculos, incluidos los revocados
     */
    totalElements?: number;
    /**
     * Cantidad total de paginas disponibles
     */
    totalPages?: number;
}